+++ minifica.api/CHANGELOG.md
@@ -1,5 +1,8 @@
 ﻿### Changelog
 
+## [DevOps-10] - [2023-06-05] - Alfredo Ferretis
+* Created CRUD entity Profile.
+
 ## [DevOps-9] - [2023-06-05] - Alfredo Ferretis
 * Created CRUD entity Action.